Update to match new format
Names are now separated by a comma instead of newline.
Sometimes collaborations have a collaboration name followed by a colon which are separated by a comma as well. "collaboration, :, actual name" So the order of checks in author match was swapped to short circut the colon which is only a single character.